refactor(tui): time_entry のフィルタ modal も同様にパッケージへまとめる
issues タブと同じ構成に揃え、タブ固有のレイアウトを app.py から追い出す。

- フィルタ modal の Float 構築と描画を filter_modal.py へ分離する
- time_entries タブ関連のファイルを tui/time_entry/ へ移動し、テストも同じ
  構成に揃える

app.py は 1002 行から 967 行になった。挙動の変更はない。
