+++ src/redi/tui/time_entry/time_entry_tab.py
