+++ tests/unit/tui/time_entry/test_time_entry_tab.py
@@ -1,7 +1,7 @@
 from typing import cast
 
 from redi.api.time_entry import TimeEntry
-from redi.tui import time_entry_tab
+from redi.tui.time_entry import time_entry_tab
 from redi.tui.state import TuiState
